Cleanup SPIs, move ActivationStore to whisk.core.database. (#3873)
* Cleanup SPIs, move ActivationStore to whisk.core.database.

- Cleaned up some edges around SPIs, made their factory method "instance" where applicable.
- Moved the ActivationStore to "whisk.core.database".

* Adjust spi impl.
