+++ common/scala/src/main/resources/reference.conf
@@ -3,7 +3,7 @@
 
 whisk.spi {
   ArtifactStoreProvider = whisk.core.database.CouchDbStoreProvider
-  ActivationStoreProvider = whisk.core.entity.ArtifactActivationStoreProvider
+  ActivationStoreProvider = whisk.core.database.ArtifactActivationStoreProvider
   MessagingProvider = whisk.connector.kafka.KafkaMessagingProvider
   ContainerFactoryProvider = whisk.core.containerpool.docker.DockerContainerFactoryProvider
   LogStoreProvider = whisk.core.containerpool.logging.DockerToActivationLogStoreProvider
