+++ core/controller/src/main/scala/whisk/core/controller/actions/SequenceActions.scala
@@ -28,14 +28,12 @@ import scala.concurrent.duration._
 import scala.language.postfixOps
 import scala.util.Failure
 import scala.util.Success
-
 import akka.actor.ActorSystem
-
 import spray.json._
-
 import whisk.common.Logging
 import whisk.common.TransactionId
 import whisk.core.controller.WhiskServices
+import whisk.core.database.ActivationStore
 import whisk.core.entity._
 import whisk.core.entity.size.SizeInt
 import whisk.core.entity.types._
